+++ application/src/main/java/com/telivery/persistence/review/application/ReviewService.java
@@ -1,19 +1,19 @@
-package com.telivery.persistence.time;
+package com.telivery.persistence.review.application;
 
-import com.telivery.persistence.time.dao.TimeRepository;
+import com.telivery.persistence.review.dao.ReviewRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class TimeService {
+public class ReviewService {
 
-  private final TimeRepository timeRepository;
+  private final ReviewRepository reviewRepository;
 
   @Transactional(readOnly = true)
   public long countByRestaurantId(Long restaurantId) {
-    return timeRepository.countByRestaurantId(restaurantId);
+    return reviewRepository.countByRestaurantId(restaurantId);
   }
 
 }
